revision round #1 (move to examples/applications, 1 file, auto-download dataset)
